[ExoBundle] separates tool and resource listeners
move listener config to annotations

reuse old bank entrypoint
