+++ plugin/exo/Listener/Log/DisplayLogListener.php
@@ -3,10 +3,19 @@
 namespace UJM\ExoBundle\Listener\Log;
 
 use Claroline\CoreBundle\Event\Log\LogCreateDelegateViewEvent;
+use JMS\DiExtraBundle\Annotation as DI;
 use Symfony\Component\DependencyInjection\ContainerAware;
 
+/**
+ * @DI\Service("ujm.exo.exercise_listener")
+ */
 class DisplayLogListener extends ContainerAware
 {
+    /**
+     * @DI\Observe("create_log_details_resource-ujm_exercise-exercise_evaluated")
+     *
+     * @param LogCreateDelegateViewEvent $event
+     */
     public function onCreateLogDetails(LogCreateDelegateViewEvent $event)
     {
         $content = $this->container->get('templating')->render(
