+++ plugin/exo/Listener/Tool/QuestionBankListener.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace UJM\ExoBundle\Listener\Tool;
+
+use Claroline\CoreBundle\Event\DisplayToolEvent;
+use JMS\DiExtraBundle\Annotation as DI;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpKernel\HttpKernelInterface;
+
+/**
+ * @DI\Service("ujm.exo.question_bank_listener")
+ */
+class QuestionBankListener
+{
+    private $container;
+
+    /**
+     * @DI\InjectParams({
+     *     "container" = @DI\Inject("service_container")
+     * })
+     *
+     * @param ContainerInterface $container
+     */
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+    }
+
+    /**
+     * @DI\Observe("open_tool_desktop_ujm_questions")
+     *
+     * @param DisplayToolEvent $event
+     */
+    public function onDisplayDesktop(DisplayToolEvent $event)
+    {
+        $subRequest = $this->container->get('request')->duplicate([], null, [
+            '_controller' => 'UJMExoBundle:Question:index',
+        ]);
+
+        $response = $this->container->get('http_kernel')->handle($subRequest, HttpKernelInterface::SUB_REQUEST);
+
+        $event->setContent($response->getContent());
+    }
+}
